master/src/requirement.txt
asgiref==3.4.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
Django==3.0
django-3-jet==1.0.8
django-ckeditor==6.2.0
django-favicon==0.1.3
django-js-asset==1.2.2
django-otp==1.1.3
et-xmlfile==1.1.0
idna==3.3
mysql-connector-python==8.0.26
numpy==1.21.2
openpyxl==3.0.7
pandas==1.3.2
pdfminer.six==20200517
pdfplumber==0.5.28
Pillow==8.3.2
pycryptodome==3.10.1
PyMySQL==1.0.2
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
Wand==0.6.7
xlrd==2.0.1

djangorestframework==3.13.1
